fix:修改regionId类型为Long

+++ lkd-server/src/main/java/com/lkd/controller/RegionController.java
@@ -38,7 +38,7 @@ public class RegionController {
 
     //根据区域id查询该区域下的点位列表
     @GetMapping("/node/search")
-    public Result searchNode(String name, Integer regionId, Integer pageIndex, Integer pageSize) {
+    public Result searchNode(String name, Long regionId, Integer pageIndex, Integer pageSize) {
         PageBean<NodeVO> pageBean = regionService.searchNode(name,regionId, pageIndex, pageSize);
         return Result.success(pageBean);
     }
