+++ lkd-server/src/main/java/com/lkd/service/RegionService.java
@@ -16,5 +16,5 @@ public interface RegionService {
     boolean update(Region region);
 
     //根据区域id查询该区域下的点位列表
-    PageBean<NodeVO> searchNode(String name, Integer regionId, Integer pageIndex, Integer pageSize);
+    PageBean<NodeVO> searchNode(String name, Long regionId, Integer pageIndex, Integer pageSize);
 }
